feat(logo): add bitcoin shape to Easter egg and make morph system reusable
Extract pixel shape definitions and morph state machine into reusable
modules so the animation system can be used elsewhere (e.g. loading screens).

- Add pixel-shapes.ts: shared shape definitions, parseShape(), grid constants
- Add usePixelMorph.ts: custom hook for morph state machine (advance, revert, settle)
- Refactor Logo.tsx to consume the hook and shared constants
- Add bitcoin (₿) pixel art to the shape cycle
- Support variable pixel counts across shapes via automatic padding
- Cycle order: logo → space invader → bitcoin → heart → logo
